+++ numsigned.go
@@ -2,6 +2,7 @@ package ferrite
 
 import (
 	"fmt"
+	"reflect"
 	"strconv"
 
 	"github.com/dogmatiq/ferrite/maybe"
@@ -69,6 +70,23 @@ func (b SignedBuilder[T]) spec(req bool) variable.TypedSpec[T] {
 			NativeMin: b.min,
 			NativeMax: b.max,
 		},
+		variable.WithDocumentation[T](
+			variable.Documentation{
+				Summary: "Signed integer syntax",
+				Paragraphs: []string{
+					"Signed integers can only be specified using decimal notation. " +
+						"A leading positive sign (`+`) is **OPTIONAL**. " +
+						"A leading negative sign (`-`) is **REQUIRED** in order to specify a negative value.",
+					fmt.Sprintf(
+						"Internally, the `%s` variable is represented using a signed %d-bit integer type (`%s`); "+
+							"any value that overflows this data-type is invalid.",
+						b.name,
+						bitSize[T](),
+						reflect.TypeOf(T(0)).Kind(),
+					),
+				},
+			},
+		),
 	)
 	if err != nil {
 		panic(err.Error())