+++ numunsigned.go
@@ -2,6 +2,7 @@ package ferrite
 
 import (
 	"fmt"
+	"reflect"
 	"strconv"
 
 	"github.com/dogmatiq/ferrite/maybe"
@@ -69,6 +70,22 @@ func (b UnsignedBuilder[T]) spec(req bool) variable.TypedSpec[T] {
 			NativeMin: b.min,
 			NativeMax: b.max,
 		},
+		variable.WithDocumentation[T](
+			variable.Documentation{
+				Summary: "Unsigned integer syntax",
+				Paragraphs: []string{
+					"Unsigned integers can only be specified using decimal (base-10) notation. " +
+						"A leading sign (`+` or `-`) is not supported and **MUST NOT** be specified.",
+					fmt.Sprintf(
+						"Internally, the `%s` variable is represented using an unsigned %d-bit integer type (`%s`); "+
+							"any value that overflows this data-type is invalid.",
+						b.name,
+						bitSize[T](),
+						reflect.TypeOf(T(0)).Kind(),
+					),
+				},
+			},
+		),
 	)
 	if err != nil {
 		panic(err.Error())